requirements.txt
Django>=5.1,<5.2
djangorestframework==3.15.2
drf-yasg==1.21.9
django-filter==25.1
django-cors-headers==4.7.0
django-allauth==65.4.1
django-import-export==4.3.7
gunicorn==23.0.0
psycopg2-binary==2.9.9